backend/requirements.txt
# FastAPI 核心
fastapi==0.115.0
uvicorn[standard]==0.32.0

# 数据库
motor==3.3.2
pymongo==4.10.1

# 环境变量与配置
python-dotenv==1.0.1
pydantic-settings==2.6.0
email-validator==2.1.0

# HTTP 客户端（调用大模型 API）
httpx==0.27.2

# 语音识别 Whisper（可选，体积较大，按需安装）
# openai-whisper==20231117

# OCR 文字提取
pytesseract==0.3.10
Pillow==11.0.0

# 身份认证
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0

# 工具
python-multipart==0.0.12

# Word 文档生成
python-docx==1.1.0
